fetch SEO data for picture templates and clean up page data/typescript on others
