+++ src/layouts/BaseLayout.astro
@@ -18,12 +18,7 @@ import type {
 	Theme,
 } from '@lib/types';
 
-export interface SEOContent {
-	description?: string;
-	image?: string;
-	imageAlt?: string;
-	title?: string;
-};
+import type { SEOContent } from '@lib/types';
 
 export interface Props {
 	headerOverlay?: boolean;
