+++ src/pages/pictures/[year]/[month]/[uid].astro
@@ -9,11 +9,12 @@ import arrowRight from '@icons/arrow-right.svg?raw';
 import index from '@icons/index.svg?raw';
 
 // utilities
+import prismic from '@lib/prismic';
+import * as prismicHelpers from '@prismicio/helpers';
 import { arrayToSentence } from '@lib/stringHelpers';
 import { format } from 'date-fns';
 import { indexPage } from '@lib/routes';
-import * as prismicHelpers from '@prismicio/helpers';
-import prismic from '@lib/prismic';
+import { pictureDescription } from '@lib/pictureHelpers';
 
 // components
 import BlockList from '@components/BlockList.astro';
@@ -28,7 +29,7 @@ import PictureGallery from '@components/navigation/PictureGallery.astro';
 import Layout from '@layouts/BaseLayout.astro';
 
 // types
-import {
+import type {
 	DateField,
 	GroupField,
 	ImageField,
@@ -38,6 +39,8 @@ import {
 	SliceZone,
 	TitleField,
 } from '@prismicio/types';
+
+import type { SEOContent } from '@lib/types';
 
 export async function getStaticPaths() {
  	const pictures = await prismic.getAllByType('picture', {
@@ -85,7 +88,7 @@ export async function getStaticPaths() {
 		`,
 	});
 
- 	return pictures.map(({ uid, data }) => {
+ 	return pictures.map(({ uid, data, tags }) => {
 		const date = prismicHelpers.asDate(data.date_completed);
 
 		const editions = data.editions.map((edition) => {
@@ -150,9 +153,15 @@ export async function getStaticPaths() {
  				uid,
  			},
  			props: {
- 				...data,
+				...data,
 				editions,
 				nav,
+				seo: {
+					description: data.seo_description || pictureDescription({ data, tags }),
+					image: data.seo_image.url || data.cover.url,
+					imageAlt: data.seo_image.alt || data.cover.alt,
+					title: data.seo_title || data.title,
+				},
 				// TODO: remove this once edition slices are deleted
 				body: data.body.filter(({ slice_type }) => slice_type !== 'edition'),
  			},
@@ -189,6 +198,7 @@ export interface Props {
 	editions?: Edition[];
 	height?: number;
 	media?: GroupField;
+	seo: SEOContent,
 	substrate?: LinkField;
 	width?: number;
 }
@@ -202,6 +212,7 @@ const {
 	editions,
 	height,
 	media,
+	seo,
 	substrate,
 	width,
 } = Astro.props as Props;
@@ -228,7 +239,11 @@ function createMediaString() {
 }
 ---
 
-<Layout path="pictures" headerOverlay>
+<Layout
+	path="pictures"
+	headerOverlay
+	{seo}
+>
 	<!-- cover -->
 	<header class="cover">
 		<div class="cover-image">
